+++ DungeonsAndDragons-DB/Tables/Enums/ActionTypes.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[ActionTypes]
 (
-    ActionTypeId INT IDENTITY(1,1)
+    ActionTypeId TINYINT IDENTITY(1,1)
 		CONSTRAINT ActionTypes_ActionTypeId_PK PRIMARY KEY,
     Name VARCHAR(50) NOT NULL,
 	FlavourText VARCHAR(MAX)
