+++ DungeonsAndDragons-DB/Tables/Enums/ArmourWeightClasses.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[ArmourWeightClasses]
 (
-    ArmourWeightClassId INT IDENTITY(1,1) 
+    ArmourWeightClassId TINYINT IDENTITY(1,1) 
 		CONSTRAINT ArmourWeightClasses_ArmourWeightClassId_PK PRIMARY KEY,
     Name VARCHAR(50) NOT NULL
 );
