+++ DungeonsAndDragons-DB/Tables/Enums/WeaponProperties.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[WeaponProperties]
 (
-    WeaponPropertyId INT IDENTITY(1,1) 
+    WeaponPropertyId TINYINT IDENTITY(1,1) 
 		CONSTRAINT WeaponProperties_WeaponPropertyId_PK PRIMARY KEY,
     Name VARCHAR(50) NOT NULL,
 	FlavourText VARCHAR(MAX)
